// src/app/shared/interfaces/product.interface.ts
export interface IProduct {
    id: string;
    catID: string;
    catName: string;
    name: string;
    sort: string;
    volume: string;
    promille: string;
    vintage: number;
    oak: string;
    color: string;
    aroma: string;
    taste: string;
    temp: string;
    price: number;
    image: string;
    count?: number;
}